Modified the retrieveTeachersList method of firebase util class:
- Decided to remove the Query method I used in previous commit to retrieve teacher records.
Used my inital way of simply retrieving all records and looping and selecting the ones that match
an input name because much easier to know when the loop is done to return to the calling Activity. This will help me handle async tasks. As a solution for the delay, I will seperate this into a seperate thread.
- Organized all teacher related activities into a seperate folder.
- Renamed the teacher related activities into fragments since all of them extend fragments now.