Fix inflater in adapter, improve date display
Inflater was uninitalized in CanceledClassAdapter, which caused a
NullPointerException when called. Inflater is now properly instantiated
in the constructor.

Additionally, the date stored in the CanceledClassDetails bean now
retrieves its data from the "pubDate" tag in the RSS XML, which contains
both the date and time of the cancelled class.