Fix bugs related to no cancellations
ListView in CanceledFragment will now be empty if no classes are
cancelled, and a message will properly be displayed to the user.
Additionally, should a teacher's name in the bean class ever be null or
empty, clicking on it in ShowCancelFragment will now display a message
to the user instead of attempting to pass bad data to another fragment.