Add onLongClickListener for cancelled class list
onLongClick will launch the "find friends in class" activity. This
activity calls the relevant FriendFinder05 API, which returns JSON. An
async task parses the JSON, and returns the result to the activity. If
no friends are found, a dialog box displays this to the user. If friends
are found, each name appears in a list. Clicking on a list object opens
the default email app, with an email composed for this friend's email
address.